+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/registry/group/DeviceGroupRegistryImplTest.java
@@ -5,7 +5,6 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
-
 package org.opendaylight.openflowplugin.impl.registry.group;
 
 import java.util.concurrent.atomic.AtomicInteger;
@@ -13,6 +12,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.group.types.rev131018.GroupId;
+import org.opendaylight.yangtools.yang.common.Uint32;
 
 /**
  * Test for {@link DeviceGroupRegistryImpl}.
@@ -26,8 +26,8 @@ public class DeviceGroupRegistryImplTest {
     @Before
     public void setUp() {
         deviceGroupRegistry = new DeviceGroupRegistryImpl();
-        groupId = new GroupId(42L);
-        groupId2 = new GroupId(84L);
+        groupId = new GroupId(Uint32.valueOf(42));
+        groupId2 = new GroupId(Uint32.valueOf(84));
         Assert.assertEquals(0, deviceGroupRegistry.getAllGroupIds().size());
         deviceGroupRegistry.store(groupId);
         Assert.assertEquals(1, deviceGroupRegistry.getAllGroupIds().size());
